第8次提交，已经可以完整的从数据库提取数据，以json的形式发送给对方

+++ src/main/java/javabean/HosttableEntity.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 @Entity
 @Table(name = "hosttable", schema = "objectdetect", catalog = "")
 public class HosttableEntity {
-    private byte time;
+    private int time;
     private Long unknown;
     private Long broadbandRouter;
     private Long wap;
@@ -15,11 +15,11 @@ public class HosttableEntity {
 
     @Id
     @Column(name = "Time")
-    public byte getTime() {
+    public int getTime() {
         return time;
     }
 
-    public void setTime(byte time) {
+    public void setTime(int time) {
         this.time = time;
     }
